+++ sentry-android-core/src/main/java/io/sentry/android/core/SentryPerformanceProvider.java
@@ -165,6 +165,7 @@ public final class SentryPerformanceProvider extends EmptySecureContentProvider 
       return;
     }
 
+    final @NotNull SentryExecutorService startupExecutorService = new SentryExecutorService();
     final @NotNull IContinuousProfiler appStartContinuousProfiler =
         new AndroidContinuousProfiler(
             buildInfoProvider,
@@ -173,7 +174,7 @@ public final class SentryPerformanceProvider extends EmptySecureContentProvider 
             logger,
             profilingOptions.getProfilingTracesDirPath(),
             profilingOptions.getProfilingTracesHz(),
-            new SentryExecutorService());
+            () -> startupExecutorService);
     appStartMetrics.setAppStartProfiler(null);
     appStartMetrics.setAppStartContinuousProfiler(appStartContinuousProfiler);
     logger.log(SentryLevel.DEBUG, "App start continuous profiling started.");
@@ -203,6 +204,7 @@ public final class SentryPerformanceProvider extends EmptySecureContentProvider 
       return;
     }
 
+    final @NotNull SentryExecutorService executorService = new SentryExecutorService();
     final @NotNull ITransactionProfiler appStartProfiler =
         new AndroidTransactionProfiler(
             context,
@@ -212,7 +214,7 @@ public final class SentryPerformanceProvider extends EmptySecureContentProvider 
             profilingOptions.getProfilingTracesDirPath(),
             profilingOptions.isProfilingEnabled(),
             profilingOptions.getProfilingTracesHz(),
-            new SentryExecutorService());
+            () -> executorService);
     appStartMetrics.setAppStartContinuousProfiler(null);
     appStartMetrics.setAppStartProfiler(appStartProfiler);
     logger.log(SentryLevel.DEBUG, "App start profiling started.");